Fix a float32_attention_logits bugs

+++ transformer_engine/jax/flax/transformer.py
@@ -178,7 +178,8 @@ def core_attention(query: Array,
 
     attn_weights = Softmax(softmax_type=softmax_type,
                            scale_factor=fused_scale_factor,
-                           sharding_type=softmax_sharding_type)(attn_weights, mask, bias)
+                           sharding_type=softmax_sharding_type)(attn_weights, mask,
+                                                                bias).astype(dtype)
 
     if not deterministic and dropout_rate > 0.:
         keep_prob = 1.0 - dropout_rate
